Console has become CustomConsole. New Console type replaces it which integrates TextSurface, IConsole, and ITextSurfaceRenderer into a single easy to use class!
